Various fixes for warnings

+++ cis_interface/communication/ZMQComm.h
@@ -226,14 +226,14 @@ int zmq_comm_recv(const comm_t x, char **data, const size_t len,
     } else {
       cislog_error("zmq_comm_recv(%s): buffer (%d bytes) is not large enough for message (%d bytes)",
 		   x.name, len, len_recv);
-      return -(len_recv + 1);
+      return -((int)len_recv + 1);
     }
   }
   memcpy(*data, zframe_data(out), len_recv + 1);
   (*data)[len_recv] = '\0';
   zframe_destroy(&out);
   cislog_debug("zmq_comm_recv(%s): returning %d", x.name, len_recv);
-  return len_recv;
+  return (int)len_recv;
 };
